Many hours wasted
it turned out that leaflet doesnt want to get rendered on server side ant ir caused errors that prevented vercel form compiling it. i moved it to another component and and imported it with turned off ssr but it didnt work. after searching forums and seeing others with same problems i changed the solution to lazy rendering but to no avail. Even more hours later i stumbled upon information that off that i did worked up to nextjs v14 but on nextjs v15 there is a little problem and it will work but i have to set reactStrictMode: false,